+++ docs/source/whats_new.rst
@@ -10,6 +10,10 @@ What's new?
    * - Version
      - Date
      -
+   * - Next
+     - In progress
+     - - :func:`pymaid.get_annotation_graph` deprecated in favour of the new
+         :func:`pymaid.get_entity_graph`.
    * - 2.1.0
      - 04/04/22
      - With this release we mainly follow some renamed functions in ``navis`` but